requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
discord.py==1.6.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
kiwisolver
matplotlib==3.3.4
multidict==5.1.0
mypy-extensions==0.4.3
numpy
packaging==20.9
pandas
pathspec
Pillow
protobuf
pyasn1
pyasn1-modules
pyparsing
python-dateutil
python-dotenv
pytz
regex
requests
rsa
six
tabulate
toml
typed-ast
typing-extensions
uritemplate
urllib3
yarl
